Extract shared control-plane test support
