+++ control-plane/tests/test_wazuh_adapter.py
@@ -1,7 +1,6 @@
 from __future__ import annotations
 
 from datetime import datetime, timezone
-import json
 import pathlib
 import sys
 import unittest
@@ -10,15 +9,15 @@ import unittest
 CONTROL_PLANE_ROOT = pathlib.Path(__file__).resolve().parents[1]
 if str(CONTROL_PLANE_ROOT) not in sys.path:
     sys.path.insert(0, str(CONTROL_PLANE_ROOT))
+TESTS_ROOT = pathlib.Path(__file__).resolve().parent
+if str(TESTS_ROOT) not in sys.path:
+    sys.path.insert(0, str(TESTS_ROOT))
 
 from aegisops_control_plane.adapters.wazuh import WazuhAlertAdapter
+from support.fixtures import load_wazuh_fixture
 
 
-FIXTURES_ROOT = pathlib.Path(__file__).resolve().parent / "fixtures" / "wazuh"
-
-
-def _load_fixture(name: str) -> dict[str, object]:
-    return json.loads((FIXTURES_ROOT / name).read_text(encoding="utf-8"))
+_load_fixture = load_wazuh_fixture
 
 
 class WazuhAlertAdapterTests(unittest.TestCase):
